AMD64: unit test progress + update + bugfixes. UNISIM::UTIL::SYMBOLIC: c++11 leverage + new expression comparison architecture
Former-commit-id: 84b9706f35944a3fa0772b9fc7c822a00faebdfd

+++ unisim_lib/unisim/util/symbolic/identifier.hh
@@ -51,7 +51,7 @@ namespace symbolic {
         if (strcmp(self.c_str(), src) == 0)
           return;
     }
-    intptr_t cmp( T rhs ) const { return int(static_cast<T const*>(this)->code) - int(rhs.code); }
+    int cmp( T rhs ) const { return int(static_cast<T const*>(this)->code) - int(rhs.code); }
     T operator + ( int offset ) const { return T( typename T::Code(int(static_cast<T const*>(this)->code) + offset) ); }
     bool next()
     {
